refactor(planning_data_analyzer): organize EPDMS metric files (#421)
* refactor(planning_data_analyzer): organize EPDMS metric files


* style(pre-commit): autofix

* refactor(planning_data_analyzer): use source-root-relative includes

* style(pre-commit): autofix

---------
