+++ planning/autoware_planning_data_analyzer/CMakeLists.txt
@@ -9,22 +9,27 @@ ament_auto_add_library(${PROJECT_NAME} SHARED
   src/base_evaluator.cpp
   src/open_loop_evaluator.cpp
   src/metrics/deviation_metrics.cpp
-  src/metrics/epdms_aggregation.cpp
-  src/metrics/ego_progress.cpp
-  src/metrics/driving_direction_compliance.cpp
-  src/metrics/history_comfort.cpp
-  src/metrics/lane_keeping.cpp
-  src/metrics/metric_utils.cpp
-  src/metrics/drivable_area_compliance.cpp
-  src/metrics/extended_comfort.cpp
-  src/metrics/ttc_within_bound.cpp
-  src/metrics/no_at_fault_collision.cpp
-  src/metrics/traffic_light_compliance.cpp
+  src/metrics/epdms/aggregation/epdms_aggregation.cpp
+  src/metrics/epdms/subscores/ego_progress.cpp
+  src/metrics/epdms/subscores/driving_direction_compliance.cpp
+  src/metrics/epdms/subscores/history_comfort.cpp
+  src/metrics/epdms/subscores/lane_keeping.cpp
+  src/metrics/geometry/metric_utils.cpp
+  src/metrics/epdms/subscores/drivable_area_compliance.cpp
+  src/metrics/epdms/subscores/extended_comfort.cpp
+  src/metrics/epdms/subscores/ttc_within_bound.cpp
+  src/metrics/epdms/subscores/no_at_fault_collision.cpp
+  src/metrics/epdms/subscores/traffic_light_compliance.cpp
   src/metrics/trajectory_metrics.cpp
   src/utils/bag_utils.cpp
   src/utils/json_utils.cpp
   src/utils/path_utils.cpp
   src/autoware_planning_data_analyzer_node.cpp
+)
+
+target_include_directories(${PROJECT_NAME}
+  PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}/src
 )
 
 rclcpp_components_register_node(${PROJECT_NAME}
@@ -48,6 +53,12 @@ if(BUILD_TESTING)
     test/metrics/test_no_at_fault_collision.cpp
     test/metrics/test_traffic_light_compliance.cpp
   )
+
+  target_include_directories(test_metrics
+    PRIVATE
+      ${CMAKE_CURRENT_SOURCE_DIR}/src
+  )
+
   target_link_libraries(test_metrics ${PROJECT_NAME})
 
   ament_add_gtest(test_offline_evaluation
@@ -55,6 +66,12 @@ if(BUILD_TESTING)
     test/test_replay_evaluation.cpp
     test/test_open_loop_gt_source_mode.cpp
   )
+
+  target_include_directories(test_offline_evaluation
+    PRIVATE
+      ${CMAKE_CURRENT_SOURCE_DIR}/src
+  )
+
   target_link_libraries(test_offline_evaluation ${PROJECT_NAME})
   ament_target_dependencies(test_offline_evaluation autoware_test_utils)
 endif()
