+++ planning/autoware_planning_data_analyzer/src/metrics/epdms/subscores/ego_progress.cpp
@@ -14,7 +14,7 @@
 
 #include "ego_progress.hpp"
 
-#include "metric_utils.hpp"
+#include "metrics/geometry/metric_utils.hpp"
 
 #include <autoware/lanelet2_utils/geometry.hpp>
