+++ planning/autoware_planning_data_analyzer/src/metrics/epdms/subscores/extended_comfort.cpp
@@ -14,7 +14,7 @@
 
 #include "extended_comfort.hpp"
 
-#include "metric_utils.hpp"
+#include "metrics/geometry/metric_utils.hpp"
 
 #include <autoware_utils_math/normalization.hpp>
