--- planning/autoware_planning_data_analyzer/src/metrics/no_at_fault_collision.hpp
@@ -1,48 +1,0 @@
-// Copyright 2026 TIER IV, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef METRICS__NO_AT_FAULT_COLLISION_HPP_
-#define METRICS__NO_AT_FAULT_COLLISION_HPP_
-
-#include "../data_types.hpp"
-
-#include <autoware/route_handler/route_handler.hpp>
-#include <autoware_vehicle_info_utils/vehicle_info.hpp>
-
-#include <limits>
-#include <memory>
-#include <string>
-
-namespace autoware::planning_data_analyzer::metrics
-{
-
-using autoware::route_handler::RouteHandler;
-
-struct NoAtFaultCollisionResult
-{
-  double score{0.0};
-  bool available{false};
-  std::string reason{"unavailable"};
-  double infraction_time_s{std::numeric_limits<double>::infinity()};
-};
-
-NoAtFaultCollisionResult calculate_no_at_fault_collision(
-  const autoware_planning_msgs::msg::Trajectory & trajectory,
-  const std::shared_ptr<PredictedObjects> & objects,
-  const autoware::vehicle_info_utils::VehicleInfo & vehicle_info,
-  const std::shared_ptr<RouteHandler> & route_handler = nullptr);
-
-}  // namespace autoware::planning_data_analyzer::metrics
-
-#endif  // METRICS__NO_AT_FAULT_COLLISION_HPP_
