+++ planning/autoware_planning_data_analyzer/test/metrics/test_deviation_metrics.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "../../src/metrics/deviation_metrics.hpp"
+#include "metrics/deviation_metrics.hpp"
 
 #include <gtest/gtest.h>
