+++ planning/autoware_planning_data_analyzer/test/metrics/test_drivable_area_compliance.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "../../src/metrics/drivable_area_compliance.hpp"
+#include "metrics/epdms/subscores/drivable_area_compliance.hpp"
 
 #include <autoware/vehicle_info_utils/vehicle_info.hpp>
