+++ planning/autoware_planning_data_analyzer/test/metrics/test_driving_direction_compliance.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "../../src/metrics/driving_direction_compliance.hpp"
+#include "metrics/epdms/subscores/driving_direction_compliance.hpp"
 
 #include <gtest/gtest.h>
