+++ planning/autoware_planning_data_analyzer/test/metrics/test_lane_keeping.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "../../src/metrics/lane_keeping.hpp"
+#include "metrics/epdms/subscores/lane_keeping.hpp"
 
 #include <gtest/gtest.h>
