+++ planning/autoware_planning_data_analyzer/test/metrics/test_ttc_within_bound.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "../../src/metrics/ttc_within_bound.hpp"
+#include "metrics/epdms/subscores/ttc_within_bound.hpp"
 
 #include <autoware_utils_geometry/geometry.hpp>
 #include <autoware_vehicle_info_utils/vehicle_info.hpp>
